refactor(neon_talk): Merge message widgets and utilities
