+++ packages/neon/neon_talk/lib/src/pages/main.dart
@@ -5,7 +5,7 @@ import 'package:neon_framework/blocs.dart';
 import 'package:neon_framework/utils.dart';
 import 'package:neon_framework/widgets.dart';
 import 'package:neon_talk/src/blocs/talk.dart';
-import 'package:neon_talk/src/widgets/message_preview.dart';
+import 'package:neon_talk/src/widgets/message.dart';
 import 'package:neon_talk/src/widgets/room_avatar.dart';
 import 'package:neon_talk/src/widgets/unread_indicator.dart';
 import 'package:nextcloud/spreed.dart' as spreed;
